backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
svgpathtools
torch
matplotlib
pandas
scikit-learn
onnx
onnxscript